refactor(api): correct typos (target and not training)
